third_party/blink/web_tests/fast/dom/shadow/inner-scope-important-wins-expected.txt

Outer scope !important rules wins with higher specificity (v0 bug compatibility).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(target).color is "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
